Shubham Jain 1 27/01/2022

+++ app/src/main/java/com/heena/supplier/fragments/AllServicesListing.kt
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.WindowManager
 import android.view.animation.AlphaAnimation
+import androidx.appcompat.app.AlertDialog
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.heena.supplier.R
@@ -17,7 +18,9 @@ import com.heena.supplier.models.Service
 import com.heena.supplier.models.ServiceListingResponse
 import com.heena.supplier.utils.LogUtils
 import com.heena.supplier.utils.SharedPreferenceUtility
+import com.heena.supplier.utils.Utility
 import com.heena.supplier.utils.Utility.Companion.apiInterface
+import com.heena.supplier.utils.Utility.Companion.setSafeOnClickListener
 import kotlinx.android.synthetic.main.activity_home2.*
 import kotlinx.android.synthetic.main.fragment_all_services_listing.view.*
 import retrofit2.Call
@@ -49,10 +52,20 @@ class AllServicesListing : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
         mView = inflater.inflate(R.layout.fragment_all_services_listing, container, false)
-        requireActivity().iv_back.setOnClickListener {
+        Utility.changeLanguage(
+            requireContext(),
+            SharedPreferenceUtility.getInstance().get(SharedPreferenceUtility.SelectedLang, "")
+        )
+        requireActivity().iv_back.setSafeOnClickListener {
             requireActivity().iv_back.startAnimation(AlphaAnimation(1F,0.5F))
             SharedPreferenceUtility.getInstance().hideSoftKeyBoard(requireContext(), requireActivity().iv_back)
             findNavController().popBackStack()
+        }
+
+        requireActivity().iv_notification.setSafeOnClickListener {
+            requireActivity().iv_notification.startAnimation(AlphaAnimation(1F,0.5F))
+            SharedPreferenceUtility.getInstance().hideSoftKeyBoard(requireContext(), requireActivity().iv_notification)
+            findNavController().navigate(R.id.notificationsFragment)
         }
 
         getServices(false)
@@ -81,6 +94,7 @@ class AllServicesListing : Fragment() {
                 }
                 if (response.isSuccessful){
                     if (response.body()!!.status==1){
+                        serviceslisting.clear()
                         serviceslisting = response.body()!!.service as ArrayList<Service>
                         if (serviceslisting.size==0){
                             mView!!.noServicesView.visibility = View.VISIBLE
@@ -101,16 +115,38 @@ class AllServicesListing : Fragment() {
 
                             override fun onServiceDele(position: Int) {
                                 val service_id = serviceslisting.get(position).service_id
-                                deleteServices(service_id!!, position)
+                                val deleteServiceDialog = AlertDialog.Builder(requireContext())
+                                deleteServiceDialog.setCancelable(false)
+                                deleteServiceDialog.setTitle(requireContext().getString(R.string.delete_service))
+                                deleteServiceDialog.setMessage(requireContext().getString(R.string.are_you_sure_you_want_to_delete_the_service))
+                                deleteServiceDialog.setPositiveButton(requireContext().getString(R.string.delete)
+                                ) { dialog, _ ->
+                                    deleteServices(service_id!!, position)
+                                    dialog!!.dismiss()
+                                }
+                                deleteServiceDialog.setNegativeButton(requireContext().getString(R.string.cancel)
+                                ) { dialog, _ -> dialog!!.cancel() }
+                                deleteServiceDialog.show()
                             }
 
                             override fun onServiceEdit(position: Int) {
-                                val service_id = serviceslisting.get(position).service_id
-                                val status = "edit"
-                                val bundle = Bundle()
-                                bundle.putInt("service_id", service_id!!)
-                                bundle.putString("status", status)
-                                findNavController().navigate(R.id.action_myServicesFragment_to_addNewFeaturedFragment, bundle)
+                                val updateServiceDialog = AlertDialog.Builder(requireContext())
+                                updateServiceDialog.setCancelable(false)
+                                updateServiceDialog.setTitle(requireContext().getString(R.string.update_service))
+                                updateServiceDialog.setMessage(requireContext().getString(R.string.would_you_like_to_update_your_service_details))
+                                updateServiceDialog.setPositiveButton(requireContext().getString(R.string.yes)
+                                ) { dialog, _ ->
+                                    val service_id = serviceslisting.get(position).service_id
+                                    val status = "edit"
+                                    val bundle = Bundle()
+                                    bundle.putInt("service_id", service_id!!)
+                                    bundle.putString("status", status)
+                                    findNavController().navigate(R.id.action_myServicesFragment_to_addNewFeaturedFragment, bundle)
+                                    dialog!!.dismiss()
+                                }
+                                updateServiceDialog.setNegativeButton(requireContext().getString(R.string.no)
+                                ) { dialog, _ -> dialog!!.cancel() }
+                                updateServiceDialog.show()
                             }
 
                         })
@@ -123,7 +159,6 @@ class AllServicesListing : Fragment() {
                 }else{
                     mView!!.noServicesView.visibility = View.VISIBLE
                     mView!!.rvServicesList.visibility = View.GONE
-                    LogUtils.shortToast(requireContext(), getString(R.string.response_isnt_successful))
                 }
             }
 
@@ -152,6 +187,14 @@ class AllServicesListing : Fragment() {
                         serviceslisting.removeAt(position)
                         if (mView!!.rvServicesList.adapter != null) {
                             mView!!.rvServicesList.adapter!!.notifyDataSetChanged()
+                        }
+
+                        if (serviceslisting.size==0){
+                            mView!!.noServicesView.visibility = View.VISIBLE
+                            mView!!.rvServicesList.visibility = View.GONE
+                        }else{
+                            mView!!.noServicesView.visibility = View.GONE
+                            mView!!.rvServicesList.visibility = View.VISIBLE
                         }
                         LogUtils.longToast(requireContext(), response.body()!!.message)
                     }
